+++ CHANGES.md
@@ -8,6 +8,12 @@ Version 0.11.3
 
 To be released.
 
+ -  Fixed a bug where use of `Federation.setInboxDispatcher()` after
+    `Federation.setInboxListeners()` had caused a `RouterError` to be
+    thrown even if the paths match.  [[#101] by Fabien O'Carroll]
+
+[#101]: https://github.com/dahlia/fedify/pull/101
+
 
 Version 0.11.2
 --------------